selenium-exercise-1/pyproject.toml
[tool.poetry]
name = "pythonselenium"
version = "0.1.0"
description = "Poetry for PythonSelenium"
authors = ["Ganesan Pugazhenthi, Thamizhenthi <ccaetga@ucl.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
allure-behave = "^2.13.2"
allure-python-commons = "^2.13.2"
annotated-types = "^0.6.0"
argcomplete = "^3.2.2"
asyncio = "^3.4.3"
attrs = "^23.2.0"
behave = "^1.2.6"
certifi = "^2024.2.2"
cffi = "^1.16.0"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
colorama = "^0.4.6"
greenlet = "^3.0.3"
h11 = "^0.14.0"
idna = "^3.6"
iniconfig = "^2.0.0"
outcome = "^1.3.0.post0"
packaging = "^23.2"
parse = "^1.20.1"
parse-type = "^0.6.2"
pipx = "^1.4.3"
platformdirs = "^4.2.0"
playwright = "^1.41.2"
pluggy = "^1.4.0"
pycparser = "^2.21"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
pydantic_core = "^2.16.3"
pyee = "^11.0.1"
PySocks = "^1.7.1"
pytest = "^8.0.0"
pytest-base-url = "^2.1.0"
pytest-playwright = "^0.4.4"
python-box= "^7.1.1"
python-dotenv = "^1.0.1"
python-slugify = "^8.0.3"
PyYAML = "^6.0.1"
requests = "^2.31.0"
selenium = "^4.17.2"
setuptools = "^69.1.1"
six = "^1.16.0"
sniffio = "^1.3.0"
sortedcontainers = "^2.4.0"
text-unidecode = "^1.3"
trio = "^0.24.0"
trio-websocket = "^0.11.1"
typing_extensions = "^4.9.0"
urllib3 = "^2.2.0"
userpath = "^1.9.1"
webdriver-manager = "^4.0.1"
wsproto = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
